+++ Wiki/Release-Notes.md
@@ -4,9 +4,19 @@ This page contains the different release details for Tournament of Teams,
 
 | Version | Release Date |
 |----|----|
+| 1.1 | Jul 28, 2022 |
 | 1.0 | May 17, 2022 |
 
 ## Release notes
+
+### 1.1 (Jul 28, 2022)
+
+Below improvements released,
+
+- New Feature: Ability to create multiple tournaments using excel template.
+- New Feature: Ability to start multiple tournaments at the same time.
+- New Feature: Local language support (translations) available for 12 languages.
+- Mobile responsive UI implemented.
 
 ### 1.0 (May 17, 2022)